@@ -1,6 +1,6 @@
 {
-  "name": "perfecthouse",
-  "version": "1.0.0",
+  "name": "frontend",
+  "version": "1.2.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -11,51 +11,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.1.2",
-    "@angular/cdk": "^15.1.2",
-    "@angular/common": "^15.1.2",
-    "@angular/compiler": "^15.1.2",
-    "@angular/core": "^15.1.2",
-    "@angular/forms": "^15.1.2",
-    "@angular/http": "^7.2.16",
-    "@angular/material": "^15.1.2",
-    "@angular/platform-browser": "^15.1.2",
-    "@angular/platform-browser-dynamic": "^15.1.2",
-    "@angular/router": "^15.1.2",
-    "@ngrx/effects": "^15.2.1",
-    "@ngrx/router-store": "^15.2.1",
-    "@ngrx/store": "^15.2.1",
-    "@ngrx/store-devtools": "^15.2.1",
+    "@angular/animations": "^17.1.3",
+    "@angular/cdk": "^17.1.2",
+    "@angular/common": "^17.1.3",
+    "@angular/compiler": "^17.1.3",
+    "@angular/core": "^17.1.3",
+    "@angular/forms": "^17.1.3",
+    "@angular/material": "^17.1.2",
+    "@angular/platform-browser": "^17.1.3",
+    "@angular/platform-browser-dynamic": "^17.1.3",
+    "@angular/router": "^17.1.3",
+    "@ngrx/effects": "^17.1.0",
+    "@ngrx/router-store": "^17.1.0",
+    "@ngrx/store": "^17.1.0",
+    "@ngrx/store-devtools": "^17.1.0",
     "hammerjs": "^2.0.8",
-    "ng": "^0.0.0",
-    "ngrx": "0.0.1-security",
     "primeicons": "^6.0.1",
-    "primeng": "^15.1.1",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.5.0",
-    "zone.js": "~0.12.0"
+    "primeng": "^17.6.0",
+    "rxjs": "^7.8.1",
+    "tslib": "^2.6.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.2",
-    "@angular-eslint/builder": "15.2.0",
-    "@angular-eslint/eslint-plugin": "15.2.0",
-    "@angular-eslint/eslint-plugin-template": "15.2.0",
-    "@angular-eslint/schematics": "15.2.0",
-    "@angular-eslint/template-parser": "15.2.0",
-    "@angular/cli": "^15.1.2",
-    "@angular/compiler-cli": "^15.1.2",
-    "@angular/language-service": "^15.1.2",
-    "@types/node": "^18.11.18",
-    "@typescript-eslint/eslint-plugin": "5.48.1",
-    "@typescript-eslint/eslint-plugin-tslint": "^5.49.0",
-    "@typescript-eslint/parser": "5.48.1",
-    "eslint": "^8.31.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jsdoc": "^39.7.0",
+    "@angular-devkit/build-angular": "^17.1.3",
+    "@angular-eslint/builder": "^17.2.1",
+    "@angular-eslint/eslint-plugin": "^17.2.1",
+    "@angular-eslint/eslint-plugin-template": "^17.2.1",
+    "@angular-eslint/schematics": "^17.2.1",
+    "@angular-eslint/template-parser": "^17.2.1",
+    "@angular/cli": "^17.1.3",
+    "@angular/compiler-cli": "^17.1.3",
+    "@angular/language-service": "^17.1.3",
+    "@types/node": "^18.19.15",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsdoc": "^48.0.6",
     "eslint-plugin-prefer-arrow": "^1.2.3",
-    "prettier": "^2.8.3",
-    "ts-node": "~10.9.1",
-    "typescript": "^4.9.4"
+    "prettier": "^3.2.5",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   }
 }
