@@ -1,38 +1,39 @@
 import {
-  ChangeDetectorRef,
   Component,
   HostListener,
   OnDestroy,
   OnInit,
-} from "@angular/core";
-import { WindowSizeDetector } from "src/app/shared/services/window-size-detector.service";
-import { trigger, style, animate, transition } from "@angular/animations";
-import { PrimeNGConfig } from "primeng/api";
-import { ABOUT_US_LINKS, ALL_LINKS } from "./header/menu-links";
-import { Title } from "@angular/platform-browser";
-import { ActivatedRoute, NavigationEnd, Router } from "@angular/router";
-import { interval, Subscription, of } from "rxjs";
-import { delayWhen } from "rxjs/operators";
-import { OffersStateManager } from "./offers/state-management/state-manager.service";
+  ChangeDetectorRef,
+} from '@angular/core';
+import { WindowSizeDetector } from 'src/app/shared/services/window-size-detector.service';
+import { trigger, style, animate, transition } from '@angular/animations';
+import { PrimeNGConfig } from 'primeng/api';
+import { ABOUT_US_LINKS, ALL_LINKS } from './header/menu-links';
+import { Title } from '@angular/platform-browser';
+import { ActivatedRoute, NavigationEnd, Router } from '@angular/router';
+import { interval, Subscription, of } from 'rxjs';
+import { delayWhen } from 'rxjs/operators';
+import { OffersStateManager } from './offers/state-management/state-manager.service';
+import { SnackbarService } from './shared/services/snackbar.service';
 
-const sideNavId = "sideNavToggleButton";
+const sideNavId = 'sideNavToggleButton';
 
 @Component({
-  selector: "perfect-root",
-  templateUrl: "./app.component.html",
-  styleUrls: ["./app.component.scss"],
+  selector: 'perfect-root',
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.scss'],
   animations: [
-    trigger("openCloseSideNavAnimation", [
-      transition(":enter", [
+    trigger('openCloseSideNavAnimation', [
+      transition(':enter', [
         style({ height: 0, width: 0, opacity: 0 }),
         animate(
-          "0.1s ease-out",
-          style({ height: 225, width: 200, opacity: 1 })
+          '0.1s ease-out',
+          style({ height: 225, width: 200, opacity: 1 }),
         ),
       ]),
-      transition(":leave", [
+      transition(':leave', [
         style({ height: 225, width: 200, opacity: 1 }),
-        animate("0.1s ease-in", style({ height: 0, width: 0, opacity: 0 })),
+        animate('0.1s ease-in', style({ height: 0, width: 0, opacity: 0 })),
       ]),
     ]),
   ],
@@ -46,28 +47,51 @@ export class AppComponent implements OnInit, OnDestroy {
   aboutUsLinks = ABOUT_US_LINKS;
   // Give time for router outlet to render its contents.
   isLoadingWithSlightDelay$ = this.offersStateManager.isLoading$.pipe(
-    delayWhen((isLoading) => (isLoading ? of(isLoading) : interval(250)))
+    delayWhen((isLoading) => (isLoading ? of(isLoading) : interval(250))),
   );
+  isSnackbarVisible = false;
+  snackbarContent = '';
 
   constructor(
     readonly windowSizeDetector: WindowSizeDetector,
     readonly offersStateManager: OffersStateManager,
     private readonly titleService: Title,
-    private readonly router: Router,
+    private readonly primengConfig: PrimeNGConfig,
+    private readonly snackbarService: SnackbarService,
     private readonly changeDetector: ChangeDetectorRef,
-    private readonly primengConfig: PrimeNGConfig
+    router: Router,
   ) {
     this.subscription = router.events.subscribe((event) => {
       if (event instanceof NavigationEnd) {
         if (!this.isUrlTitleComputedInComponent(event.url)) {
           const title = this.getTitle(
             router.routerState,
-            router.routerState.root
-          ).join(" - ");
+            router.routerState.root,
+          ).join(' - ');
           this.titleService.setTitle(title);
         }
       }
     });
+    this.subscription.add(
+      this.snackbarService.open$.subscribe((message) => {
+        this.openSnackbar(message);
+        setTimeout(() => {
+          this.closeSnackbar();
+        }, 3000);
+      }),
+    );
+  }
+
+  private openSnackbar(message: string) {
+    this.snackbarContent = message;
+    this.isSnackbarVisible = true;
+    this.changeDetector.detectChanges();
+  }
+
+  private closeSnackbar() {
+    this.snackbarContent = '';
+    this.isSnackbarVisible = false;
+    this.changeDetector.detectChanges();
   }
 
   ngOnInit() {
@@ -81,7 +105,7 @@ export class AppComponent implements OnInit, OnDestroy {
 
   isUrlTitleComputedInComponent(url: string): boolean {
     return (
-      (url.startsWith("/ludzie/") && url.length > 8) || url.includes("oferta")
+      (url.startsWith('/ludzie/') && url.length > 8) || url.includes('oferta')
     );
   }
 
@@ -97,7 +121,7 @@ export class AppComponent implements OnInit, OnDestroy {
     return data;
   }
 
-  @HostListener("window:resize", ["$event"])
+  @HostListener('window:resize', ['$event'])
   onResize() {
     this.windowSizeDetector.windowSizeChanged(window.innerWidth);
   }
@@ -112,13 +136,16 @@ export class AppComponent implements OnInit, OnDestroy {
     }
   }
 
-  @HostListener("document:click", ["$event"])
+  @HostListener('document:click', ['$event'])
   onClick(event: any) {
     if (
       this.windowSizeDetector.isWindowSmallerThanDesktopSmall &&
-      this.isSideMenuVisible && !!event?.target
+      this.isSideMenuVisible &&
+      !!event?.target
     ) {
-      const isSideNavButtonClicked = event.target.id === sideNavId || event.target.offsetParent?.id === sideNavId;
+      const isSideNavButtonClicked =
+        event.target.id === sideNavId ||
+        event.target.offsetParent?.id === sideNavId;
       if (!isSideNavButtonClicked) {
         this.isSideMenuVisible = false;
       }
