@@ -3,16 +3,16 @@ import {
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   OnDestroy,
-} from "@angular/core";
-import { Subscription } from "rxjs";
-import { OffersStateManager } from "src/app/offers/state-management/state-manager.service";
-import { WindowSizeDetector } from "src/app/shared/services/window-size-detector.service";
-import { AgentsStateManager } from "../state-management/state-manager.service";
+} from '@angular/core';
+import { Subscription } from 'rxjs';
+import { OffersStateManager } from 'src/app/offers/state-management/state-manager.service';
+import { WindowSizeDetector } from 'src/app/shared/services/window-size-detector.service';
+import { AgentsStateManager } from '../state-management/state-manager.service';
 
 @Component({
-  selector: "perfect-agent-page",
-  templateUrl: "./agent-page.component.html",
-  styleUrls: ["./agent-page.component.scss"],
+  selector: 'perfect-agent-page',
+  templateUrl: './agent-page.component.html',
+  styleUrls: ['./agent-page.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AgentPageComponent implements OnDestroy {
@@ -22,12 +22,12 @@ export class AgentPageComponent implements OnDestroy {
     readonly agentsStateManager: AgentsStateManager,
     readonly offersStateManager: OffersStateManager,
     readonly windowSizeDetector: WindowSizeDetector,
-    private readonly changeDetector: ChangeDetectorRef
+    private readonly changeDetector: ChangeDetectorRef,
   ) {
     this.subscription = this.windowSizeDetector.windowSizeChanged$.subscribe(
       () => {
         this.changeDetector.detectChanges();
-      }
+      },
     );
   }
 
