@@ -1,18 +1,18 @@
-import { Injectable } from "@angular/core";
+import { Injectable } from '@angular/core';
 
-import { Agent } from "../../shared/models";
+import { Agent } from '../../shared/models';
 
 @Injectable({
-  providedIn: "root",
+  providedIn: 'root',
 })
 export class AgentsConverter {
   convertToReadableAgents(rawAgents: any[]): Agent[] {
     return rawAgents
-      .filter((agent) => agent.Nazwa !== "Biuro Perfecthouse")
+      .filter((agent) => agent.Nazwa !== 'Biuro Perfecthouse')
       .map((agent) => {
         return {
           id: agent.ID,
-          fullName: agent.Imie + " " + agent.Nazwisko,
+          fullName: agent.Imie + ' ' + agent.Nazwisko,
           position: this.computePosition(agent.Nazwa),
           photoFileName: agent.PlikFoto,
           phone: agent.Telefon,
@@ -25,6 +25,6 @@ export class AgentsConverter {
   }
 
   private computePosition(nameWithPosition: string) {
-    return nameWithPosition.split(" - ")[1];
+    return nameWithPosition.split(' - ')[1];
   }
 }
