@@ -1,32 +1,34 @@
-import { Injectable } from "@angular/core";
-import { Store } from "@ngrx/store";
-import { Observable } from "rxjs";
-import { openAgentPage } from "src/app/router/state-management/actions";
-import { Agent } from "src/app/shared/models";
-import { listAgents, loadNextAgent, loadPreviousAgent } from "./actions";
-import * as selectors from "./selectors";
+import { Injectable } from '@angular/core';
+import { Store } from '@ngrx/store';
+import { Observable } from 'rxjs';
+import { openAgentPage } from 'src/app/router/state-management/actions';
+import { Agent } from 'src/app/shared/models';
+import { listAgents, loadNextAgent, loadPreviousAgent } from './actions';
+import * as selectors from './selectors';
 
 @Injectable({
-  providedIn: "root",
+  providedIn: 'root',
 })
 export class AgentsStateManager {
   constructor(private store: Store) {
     this.store.dispatch(listAgents());
   }
 
-  agents$: Observable<Agent[]> = this.store.select(selectors.getUniqueAgents);
+  get agents$(): Observable<Agent[]> {
+    return this.store.select(selectors.getUniqueAgents);
+  }
 
-  currentAgent$: Observable<Agent> = this.store.select(
-    selectors.getCurrentAgent
-  );
+  get currentAgent$(): Observable<Agent> {
+    return this.store.select(selectors.getCurrentAgent);
+  }
 
-  isPreviousAgentAvailable$: Observable<boolean> = this.store.select(
-    selectors.getIsPreviousAgentAvailable
-  );
+  get isPreviousAgentAvailable$(): Observable<boolean> {
+    return this.store.select(selectors.getIsPreviousAgentAvailable);
+  }
 
-  isNextAgentAvailable$: Observable<boolean> = this.store.select(
-    selectors.getIsNextAgentAvailable
-  );
+  get isNextAgentAvailable$(): Observable<boolean> {
+    return this.store.select(selectors.getIsNextAgentAvailable);
+  }
 
   agentById$(id: number): Observable<Agent> {
     return this.store.select(selectors.getAgentById(id));
