@@ -1,4 +1,9 @@
-import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
+import {
+  Component,
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  OnDestroy,
+} from '@angular/core';
 import { Subscription } from 'rxjs';
 import { AgentsStateManager } from '../agents/state-management/state-manager.service';
 import { WindowSizeDetector } from '../shared/services/window-size-detector.service';
@@ -21,7 +26,7 @@ const HOUSES_PHOTO_NAMES = [
   styleUrls: ['./eco-houses.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class EcoHousesComponent {
+export class EcoHousesComponent implements OnDestroy {
   private subscription: Subscription;
   agentResponsibleForEcoHousesId = AGENT_RESPONSIBLE_FOR_MANAGEMENT_ID;
   housesPhotoUrls = HOUSES_PHOTO_NAMES.map(convertPhotoNameToUrl);
@@ -29,12 +34,12 @@ export class EcoHousesComponent {
   constructor(
     readonly windowSizeDetector: WindowSizeDetector,
     readonly agentsStateManager: AgentsStateManager,
-    private readonly changeDetector: ChangeDetectorRef
+    private readonly changeDetector: ChangeDetectorRef,
   ) {
     this.subscription = this.windowSizeDetector.windowSizeChanged$.subscribe(
       () => {
         this.changeDetector.detectChanges();
-      }
+      },
     );
   }
 
