@@ -3,44 +3,45 @@ import {
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   OnDestroy,
-} from "@angular/core";
-import { Router } from "@angular/router";
-import { Subscription } from "rxjs";
-import { WindowSizeDetector } from "../shared/services/window-size-detector.service";
+} from '@angular/core';
+import { Router } from '@angular/router';
+import { Subscription } from 'rxjs';
+import { WindowSizeDetector } from '../shared/services/window-size-detector.service';
 import {
   DEFAULT_FILTERS,
   GalleryPhoto,
   OffersFilters,
-} from "src/app/shared/models";
-import { OffersStateManager } from "../offers/state-management/state-manager.service";
+} from 'src/app/shared/models';
+import { OffersStateManager } from '../offers/state-management/state-manager.service';
+import { ContactFormType } from '../contact/contact-form/contact-form.component';
 
 const IMAGES: GalleryPhoto[] = [
   {
-    previewImageSrc: "/assets/dla_ciebie.jpg",
-    thumbnailImageSrc: "/assets/dla_ciebie.jpg",
-    title: "Sprawdź nasze oferty dla Ciebie.",
+    previewImageSrc: '/assets/dla_ciebie.jpg',
+    thumbnailImageSrc: '/assets/dla_ciebie.jpg',
+    title: 'Sprawdź nasze oferty dla Ciebie.',
   },
   {
-    previewImageSrc: "/assets/nad_morzem.jpg",
-    thumbnailImageSrc: "/assets/nad_morzem.jpg",
-    title: "Zobacz inwestycje nad morzem.",
+    previewImageSrc: '/assets/nad_morzem.jpg',
+    thumbnailImageSrc: '/assets/nad_morzem.jpg',
+    title: 'Zobacz inwestycje nad morzem.',
   },
   {
-    previewImageSrc: "/assets/po_poznansku.jpg",
-    thumbnailImageSrc: "/assets/po_poznansku.jpg",
-    title: "Sprawdź nasze oferty dla Ciebie.",
+    previewImageSrc: '/assets/po_poznansku.jpg',
+    thumbnailImageSrc: '/assets/po_poznansku.jpg',
+    title: 'Sprawdź nasze oferty dla Ciebie.',
   },
   {
-    previewImageSrc: "/assets/zarzadzanie.jpg",
-    thumbnailImageSrc: "/assets/zarzadzanie.jpg",
-    title: "Zobacz, jakie proste może być zarządzanie nieruchomościami.",
+    previewImageSrc: '/assets/zarzadzanie.jpg',
+    thumbnailImageSrc: '/assets/zarzadzanie.jpg',
+    title: 'Zobacz, jakie proste może być zarządzanie nieruchomościami.',
   },
 ];
 /** Strona główna. */
 @Component({
-  selector: "perfect-main",
-  templateUrl: "./main.component.html",
-  styleUrls: ["./main.component.scss"],
+  selector: 'perfect-main',
+  templateUrl: './main.component.html',
+  styleUrls: ['./main.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class MainComponent implements OnDestroy {
@@ -53,38 +54,38 @@ export class MainComponent implements OnDestroy {
     private readonly router: Router,
     readonly offersStateManager: OffersStateManager,
     readonly windowSizeDetector: WindowSizeDetector,
-    readonly changeDetector: ChangeDetectorRef
+    readonly changeDetector: ChangeDetectorRef,
   ) {
     this.subscription = this.windowSizeDetector.windowSizeChanged$.subscribe(
       () => {
         this.changeDetector.detectChanges();
-      }
+      },
     );
   }
 
   openUrl(image) {
-    if (image.previewImageSrc.includes("zarzadzanie")) {
-      this.router.navigate(["zarządzanie"]);
+    if (image.previewImageSrc.includes('zarzadzanie')) {
+      this.router.navigate(['zarządzanie']);
       return;
     }
-    if (image.previewImageSrc.includes("poznansku")) {
-      this.router.navigate(["wartości"]);
+    if (image.previewImageSrc.includes('poznansku')) {
+      this.router.navigate(['wartości']);
       return;
     }
 
     let params = {};
-    if (image.title.includes("morze")) {
+    if (image.title.includes('morze')) {
       params = { isByTheSea: true };
     }
-    this.router.navigate(["oferty"], { queryParams: params });
+    this.router.navigate(['oferty'], { queryParams: params });
   }
 
   loadOffer(symbol: string) {
-    this.router.navigate(["oferta", symbol]);
+    this.router.navigate(['oferta', symbol]);
   }
 
   loadOffers(filters: OffersFilters) {
-    this.router.navigate(["oferty"], {
+    this.router.navigate(['oferty'], {
       queryParams: {
         ...this.computeFiltersParameters(filters),
       },
@@ -123,4 +124,6 @@ export class MainComponent implements OnDestroy {
   ngOnDestroy() {
     this.subscription.unsubscribe();
   }
+
+  protected readonly FormType = ContactFormType;
 }
