@@ -1,11 +1,11 @@
-import { createFeatureSelector, createSelector } from "@ngrx/store";
+import { createFeatureSelector, createSelector } from '@ngrx/store';
 import {
   AVAILABLE_SORTINGS,
   Offer,
   OffersFilters,
-} from "src/app/shared/models";
-import { OffersState } from "./reducers";
-import { sortOffers } from "./sorting-helper-functions";
+} from 'src/app/shared/models';
+import { OffersState } from './reducers';
+import { sortOffers } from './sorting-helper-functions';
 import {
   computeAgentOffers,
   computeDistinctLocations,
@@ -13,65 +13,65 @@ import {
   computeNumberOfPages,
   computeOffersForCurrentPage,
   computeVoivodeshipsWithCounties,
-} from "./state-helper-functions";
-import { stateKey } from "./reducers";
-import { filterOffers } from "./filter-helper-functions";
+} from './state-helper-functions';
+import { stateKey } from './reducers';
+import { filterOffers } from './filter-helper-functions';
 
 const selectOffersState = createFeatureSelector<OffersState>(stateKey);
 
 export const getIsLoading = createSelector(
   selectOffersState,
-  (state: OffersState) => state.isLoading
+  (state: OffersState) => state.isLoading,
 );
 
 export const getIsSearching = createSelector(
   selectOffersState,
-  (state: OffersState) => state.isSearching
+  (state: OffersState) => state.isSearching,
 );
 
 export const getIsCurrentOfferLoading = createSelector(
   selectOffersState,
-  (state: OffersState) => !state.currentOffer
+  (state: OffersState) => !state.currentOffer,
 );
 
 export const getCurrentOffer = createSelector(
   selectOffersState,
-  (state: OffersState) => state.currentOffer
+  (state: OffersState) => state.currentOffer,
 );
 
 export const getAllOffers = createSelector(
   selectOffersState,
-  (state: OffersState) => state.allOffers
+  (state: OffersState) => state.allOffers,
 );
 
 export const getOffersForMainPage = createSelector(
   selectOffersState,
-  (state: OffersState) => state.mainPageOffers
+  (state: OffersState) => state.mainPageOffers,
 );
 
 export const getCurrentSearchOffers = createSelector(
   selectOffersState,
-  (state: OffersState) => state.currentSearchOffers
+  (state: OffersState) => state.currentSearchOffers,
 );
 
 export const getCurrentSearchOffersQuantity = createSelector(
   getCurrentSearchOffers,
-  (currentSearchOffers: Offer[]) => currentSearchOffers.length
+  (currentSearchOffers: Offer[]) => currentSearchOffers.length,
 );
 
 export const getPageNumber = createSelector(
   selectOffersState,
-  (state: OffersState) => state.pageNumber
+  (state: OffersState) => state.pageNumber,
 );
 
 export const getSorting = createSelector(
   selectOffersState,
-  (state: OffersState) => state.sorting
+  (state: OffersState) => state.sorting,
 );
 
 export const getFilters = createSelector(
   selectOffersState,
-  (state: OffersState) => state.filters
+  (state: OffersState) => state.filters,
 );
 
 export const getOffersForCurrentPage = createSelector(
@@ -79,7 +79,7 @@ export const getOffersForCurrentPage = createSelector(
   getPageNumber,
   (currentSearchOffers: Offer[], pageNumber: number) => {
     return computeOffersForCurrentPage(currentSearchOffers, pageNumber);
-  }
+  },
 );
 
 const getCurrentSearchOffersSortedByPriceAsc = createSelector(
@@ -89,17 +89,17 @@ const getCurrentSearchOffersSortedByPriceAsc = createSelector(
     return sortOffers(
       filterOffers(allOffers, filters),
       AVAILABLE_SORTINGS.find(
-        (sorting) => sorting.displayName === "cenie rosnąco"
-      )
+        (sorting) => sorting.displayName === 'cenie rosnąco',
+      ),
     );
-  }
+  },
 );
 
 export const getLowestPriceForCurrentSearch = createSelector(
   getCurrentSearchOffersSortedByPriceAsc,
   (offersSortedByPriceAsc: Offer[]) => {
     return offersSortedByPriceAsc.length ? offersSortedByPriceAsc[0].price : -1;
-  }
+  },
 );
 
 export const getHighestPriceForCurrentSearch = createSelector(
@@ -108,14 +108,14 @@ export const getHighestPriceForCurrentSearch = createSelector(
     return offersSortedByPriceAsc.length
       ? offersSortedByPriceAsc[offersSortedByPriceAsc.length - 1].price
       : -1;
-  }
+  },
 );
 
 export const getNumberOfPages = createSelector(
   getCurrentSearchOffers,
   (currentSearchOffers: Offer[]) => {
     return computeNumberOfPages(currentSearchOffers);
-  }
+  },
 );
 
 export const getDistinctLocations = (voivodeship: string, county: string) =>
@@ -142,14 +142,14 @@ export const getVoivodeshipsWithCounties = createSelector(
   getAllOffers,
   (currentSearchOffers: Offer[]) => {
     return computeVoivodeshipsWithCounties(currentSearchOffers);
-  }
+  },
 );
 
 export const getEstateTypesWithSubtypes = createSelector(
   getAllOffers,
   (currentSearchOffers: Offer[]) => {
     return computeEstateTypesWithSubtypes(currentSearchOffers);
-  }
+  },
 );
 
 export const getCurrentOfferIndex = createSelector(
@@ -162,14 +162,14 @@ export const getCurrentOfferIndex = createSelector(
     return currentSearchOffers
       .map((offer) => offer.symbol)
       .indexOf(currentOffer.symbol);
-  }
+  },
 );
 
 export const getIsPreviousOfferAvailable = createSelector(
   getCurrentOfferIndex,
   (currentOfferIndex: number) => {
     return currentOfferIndex > 0;
-  }
+  },
 );
 
 export const getIsNextOfferAvailable = createSelector(
@@ -180,5 +180,5 @@ export const getIsNextOfferAvailable = createSelector(
       currentOfferIndex !== -1 &&
       currentOfferIndex < currentSearchOffersQuantity - 1
     );
-  }
+  },
 );
