@@ -1,20 +1,20 @@
-import { Offer, Sorting } from "src/app/shared/models";
+import { Offer, Sorting } from 'src/app/shared/models';
 
 export const sortOffers = (offers: Offer[], sorting: Sorting): Offer[] => {
   if (sorting.isAscending) {
     return [...offers].sort((a, b) =>
-      compareByPropertyAsc(sorting.propertyName, a, b)
+      compareByPropertyAsc(sorting.propertyName, a, b),
     );
   }
   return [...offers].sort((a, b) =>
-    compareByPropertyDesc(sorting.propertyName, a, b)
+    compareByPropertyDesc(sorting.propertyName, a, b),
   );
 };
 
 const compareByPropertyAsc = (
   propertyName: string,
   offerA: Offer,
-  offerB: Offer
+  offerB: Offer,
 ): number => {
   const propertyA = getPropertyValue(offerA, propertyName);
   const propertyB = getPropertyValue(offerB, propertyName);
@@ -30,7 +30,7 @@ const compareByPropertyAsc = (
 const compareByPropertyDesc = (
   propertyName: string,
   offerA: Offer,
-  offerB: Offer
+  offerB: Offer,
 ): number => {
   const propertyA = getPropertyValue(offerA, propertyName);
   const propertyB = getPropertyValue(offerB, propertyName);
@@ -49,8 +49,8 @@ const getPropertyValue = (offer: Offer, propertyName: string) => {
     return offer[propertyName].value;
   }
   return offer[propertyName];
-}
+};
 
 export const sortAlphabetically = (a: string, b: string): number => {
-  return a.localeCompare(b, "pl");
+  return a.localeCompare(b, 'pl');
 };
