@@ -1,10 +1,10 @@
-import { Component, ChangeDetectionStrategy } from "@angular/core";
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 
 // Source: https://tobiasahlin.com/spinkit/
 @Component({
-  selector: "perfect-loading-spinner",
-  templateUrl: "./loading-spinner.component.html",
-  styleUrls: ["./loading-spinner.component.scss"],
+  selector: 'perfect-loading-spinner',
+  templateUrl: './loading-spinner.component.html',
+  styleUrls: ['./loading-spinner.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class LoadingSpinnerComponent {}
